pyChat/requirements.txt
appdirs==1.4.4
apturl==0.5.2
asn1crypto==0.24.0
Brlapi==0.6.6
certifi==2018.1.18
cffi==1.14.2
chardet==3.0.4
command-not-found==0.3
cryptography==3.0
cupshelpers==1.0
defer==1.0.6
distlib==0.3.1
distro-info===0.18ubuntu0.18.04.1
filelock==3.0.12
httplib2==0.9.2
idna==2.6
importlib-metadata==1.7.0
importlib-resources==3.0.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
pip==9.0.1
protobuf==3.0.0
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.27.0
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.0
python-dateutil==2.6.1
python-debian==0.1.32
python-dotenv==0.14.0
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
setuptools==39.0.1
simplejson==3.13.2
six==1.15.0
system-service==0.3
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==20.0.31
wadllib==1.3.2
wheel==0.30.0
xkit==0.0.0
zipp==3.1.0
zope.interface==4.3.2
